requirement.txt
asttokens==2.4.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
executing==2.0.1
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.49.0
greenlet==3.0.3
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mysql-connector==2.2.9
mysql-connector-python==8.4.0
mysqlclient==2.2.4
numba==0.59.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
SQLAlchemy==2.0.28
stack-data==0.6.3
threadpoolctl==3.3.0
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
wcwidth==0.2.13
Werkzeug==3.0.1